==> 4_Transcriptomics/bin/R_scripts/7_1_DGE_5TCvs5DC.Rmd <==
---
title: "7_1_DGE_5TCvs5DC"
author: "Veronica Reyes-Galindo"
date: "18/5/2021"
output:
  html_document:
    toc: true
    toc_float: true
    theme: "yeti"
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#### Load libraries

```{r libreries, echo= FALSE}
library(limma)
library(edgeR)
library(DESeq2)
library(ggbiplot)
library (ggplot2)
```


# Load and select data Damaged vs Tolerant 170 ppb

#### Load data in matrix

```{r Load data in matrix, echo= FALSE}
# Load data. Count table 
alldata <-read.delim("../../data/allreadsgenes.txt")

# Convert dataframe to data matrix
x<-alldata
rownames(x)<-alldata[,1] # Add rownames
x<-x[ ,2:ncol(x)] # Remove double col with names
alldata<-as.matrix(x)
rm(x) # Remove x
```

#### Select subset data Damaged vs Tolerant 170 ppb

```{r Select subset data, echo= FALSE}
DCvsTC<- subset(alldata, select = -c(DS_1, DS_2, DS_4,
                                     TS_1, TS_2, TS_5,
                                     TC17, DC47))
```

#### Add characteristics
```{r characteristics, echo= FALSE}
treatment <- c("DC","DC","DC","DC","DC",
                 "TC","TC","TC","TC","TC")
label <- c("DC_1", "DC_2","DC_3","DC_4","DC_5",
          "TC_1","TC_2","TC_3","TC_4","TC_5")
samples <-c("DC1", "DC2","DC3","DC4","DC5",
           "TC1","TC2","TC3","TC4","TC5")
targets <- data.frame(treatment,label,samples)
rownames(targets) <- label
targets
```

#### Filtering genes

```{r Filtering genes, echo= FALSE}
table(rowSums(DCvsTC)==0) #Count O in T & D with FALSE and TRUE
suma <- rowSums(DCvsTC) # Sum reads per each gen 
filtconteos <- DCvsTC[suma != 0,] #Make matrix without zero reads per gen
```
# Analyzes DGE with edgeR and DESeq2

#### edgeR

```{r edgeR, echo= FALSE}
## Clase DGEList
#Make list with counts and samples to NORMALIZATION
d <- DGEList(counts = filtconteos[,1:10], group = targets$treatment)
colnames(d) <- targets$label

## Normalization
d <- calcNormFactors(d) 
plotMDS(d, main="plotMDS DCvsTC")

## Dispersors estimate
d <- estimateCommonDisp(d,verbose=TRUE)
d <- estimateTagwiseDisp(d)
plotBCV(d, main="plotBCV DCvsTC")

## Test
et <- exactTest(d,pair=c("DC","TC"))
top<- topTags(et, n= Inf)
hist(top$table$FDR, breaks = 100, main = "Hist FDR DCvsTC")
abline(v=0.05, col="red",lwd=3)
```

#### DESeq2

```{r DESeq2, echo= FALSE}
### Class DESeqDataSet
dds <- DESeqDataSetFromMatrix(countData=filtconteos, colData= targets, design=~treatment)

### Test
dds <- DESeq(dds)
head(dds)
res <- results(dds)
head (res)
```

#### Check normalization of both methods

```{r Check normalization of both methods, echo= FALSE}
# edgeR
d$samples$norm.factors #edgeR
# DESeq2
sizeFactors(dds) #DESeq2

### Compareted dispersion values
# edgeR
# First calculated comun disspersion
#d$common.dispersion
# Second gen to gen dispersion of comun dispersion
#head(d$tagwise.dispersion)
# Choose the best of both stimation 

# DESeq2
# First calculated stimation gen to gen  
#head(mcols(dds)$dispGeneEs)
# Then through an adjustment with the average counts estimate the dispersion
#head(mcols(dds)$dispersion)

#Compare the tests, that is, the p-values and other results that each packet has calculated for each gene
# edgeR
#topTags(et, n= Inf)
# DESeq2
#res[rownames(topTags(et, n= Inf)),]
```
# VolcanoPlots 

#### Plot Log fold change

```{r Plot Log fold change, echo= FALSE}
# edgeR
########p.value=0.1
de <- decideTestsDGE(et, adjust.method = "fdr" )
head(de)
detags <- rownames(d)[as.logical(de)]
plotSmear(et, de.tags=detags, main="plotSmear de edgeR") > abline(h=0, col="red", lwd=3)

# DESeq2
plotMA(res, main="MA-plot DESeq2", ylim=c(-5,5))
```

# FDR < 0.05

#### Sort the genes according to the attached p-value they have obtained FDR < 0.05

```{r Sort FDR0.05, echo= FALSE}
# edgeR
topSig <- top[top$table$FDR < 0.05, ]
dim(topSig)
#head(topSig)
genesDEedgeR <- rownames(topSig)
head(genesDEedgeR)
topSig_export<-topSig
topSig_export$ID<-genesDEedgeR
#head(topSig_export)
write.table(topSig_export, "../../data/DGE/EdgeR_TvsD170ppb_FDR_0.05.txt", sep="\t", row.names=T)

# DESeq2
# Sort by p-valores
resOrdered <- res[order(res$padj),]
# Only DEG
xx <-res[order(res$padj,na.last=NA),] 
head(xx)
resSig2 <- xx[xx$padj < 0.05, ]
dim(resSig2)
head(resSig2)
genesDEDESeq2 <- rownames(resSig2)
genesDEDESeq2
head(resSig2)
resSig2_export<-resSig2
resSig2_export$ID<-resSig2_export
head(resSig2_export)
write.table(resSig2_export, "../../data/DGE/DESeq2_TvsD170ppb_FDR_0.05.txt", sep="\t", row.names=T)
```

# FDR < 5

#### #### Sort the genes according to the attached p-value they have obtained FDR < 5

```{r Sort FDR5, echo= FALSE}
### Sort the genes according to the attached p-value they have obtained
# EdgeR
topSig <- top[top$table$FDR < 5, ]
dim(topSig)
genesDEedgeR <- rownames(topSig)
head(genesDEedgeR)
topSig_export<-topSig
topSig_export$ID<-genesDEedgeR
#head(topSig_export)
write.table(topSig_export, "../../data/DGE/EdgeR_TvsD170ppb_FDR_5.txt", sep="\t", row.names=T)

# DESeq2
xx <-res[order(res$padj,na.last=NA),] 
resSig2 <- xx[xx$padj < 5, ]
dim(resSig2)
head(resSig2)
genesDEDESeq2 <- rownames(resSig2)
head(resSig2)
resSig2_export<-resSig2
resSig2_export$ID<-resSig2_export
head(resSig2_export)
write.table(resSig2_export, "../../data/DGE/DESeq2_TvsD170ppb_FDR_5.txt", sep="\t", row.names=T)
```

# Common DEG in edgeR vs DESeq2

#### How many common DE genes exist edgeR vs DESeq2

```{r common DE genes, echo= FALSE}
genesDEcomunes <- intersect(genesDEedgeR,genesDEDESeq2) 
head(genesDEcomunes)
str(genesDEcomunes)
```
