==> 4_Transcriptomics/bin/R_scripts/8_1_Volcanoplot_merge_paper.Rmd <==
---
title: "8_1_Volcanoplot_merge_paper"
author: "Veronica Reyes-Galindo"
date: "18/5/2021"
output:
  html_document:
    toc: true
    toc_float: true
    theme: "yeti"
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#### Load libraries

```{r libreries, echo= FALSE}
library(ggplot2)
library(ggpubr)
```

# Load data 

#### Load data

```{r Load data, echo= FALSE}
C_TvsD<-read.delim("../../data/Over_Down/GENERAL_C_TvsD_inDamaged.txt", sep="\t")
D_170Cvs87SS<-read.delim("../../data/Over_Down/GENERAL_D_170Cvs87SS_inConti.txt", sep="\t")
T_170Cvs87SS<-read.delim("../../data/Over_Down/GENERAL_T_170Cvs87SS_inConti.txt", sep="\t")
S_TvsD<-read.delim("../../data/Over_Down/GENERAL_S_TvsD_inDamaged.txt", sep="\t")
```

# Plots


color 1= purple D2 and ER

color 2= blue Only D2

color 3= yellow Only ER

## C_TvsD

Down and up in tolerant trees

```{r Plot individual C_TvsD, echo= FALSE}
ggplot(C_TvsD, aes(x=log2FoldChange_D2, y=sig_D2)) +
  geom_point(aes(colour =  color ),size =3.5)+
  scale_color_manual(values=c("#cd77ca", # purple D2 and ER
                              "#49cfd8", # blue Only D2
                              "#cdbd3a", # yellow Only ER
                              "grey",    # grey
                              "black"))+ # black
  xlab("log2 fold change")+
  ylab("-log10 (P value)")+
  theme_light(base_size = 10)+
  theme(legend.position = "none")+
  geom_hline(yintercept = -log10(0.05), colour = "black", linetype = "dashed", size = 0.25) + # Horizontal significance cut-off line.
  geom_vline(xintercept = 1, colour = "black", linetype = "dashed", size = 0.25)+  # Vertical significance cut-off line (+).
  geom_vline(xintercept = -1, colour = "black", linetype = "dashed", size = 0.25)  # Vertical significance cut-off line (+).
```

## D_170Cvs87SS

Table data to down and up in SS period, but I change plot to see in Contingency period, so is down and up in contingency 

```{r Plot individual D_170Cvs87SS, echo= FALSE}
ggplot(D_170Cvs87SS, aes(x=log2FoldChange_D2, y=sig_D2)) +
  geom_point(aes(colour =  color ),size =3.5)+
  scale_color_manual(values=c("#cd77ca", # purple D2 and ER
                              "#49cfd8", # blue Only D2
                              "grey",    # grey
                              "black"))+ # black
  xlab("log2 fold change")+
  ylab("-log10 (P value)")+
  theme_light(base_size = 10)+
  theme(legend.position = "none")+
  geom_hline(yintercept = -log10(0.05), colour = "black", linetype = "dashed", size = 0.25) + # Horizontal significance cut-off line.
  geom_vline(xintercept = 1, colour = "black", linetype = "dashed", size = 0.25)+  # Vertical significance cut-off line (+).
  geom_vline(xintercept = -1, colour = "black", linetype = "dashed", size = 0.25)  # Vertical significance cut-off line (+).
ggsave("../../outputs/8_1_VP_General_sinN_D_170Cvs87SS.tiff")
```
## T_170Cvs87SS

Table data to down and up in SS period, but I change plot to see in Contingency period, so is down and up in contingency 

```{r Plot individual T_170Cvs87SS, echo= FALSE}
ggplot(T_170Cvs87SS, aes(x=log2FoldChange_D2, y=sig_D2)) +
  geom_point(aes(colour =  color ),size =3.5)+
  scale_color_manual(values=c("#cd77ca", # purple D2 and ER
                              "#49cfd8", # blue Only D2
                              "#cdbd3a", # yellow Only ER
                              "grey",    # grey
                              "black"))+ # black
  xlab("log2 fold change")+
  ylab("-log10 (P value)")+
  theme_light(base_size = 10)+
  theme(legend.position = "none")+
  geom_hline(yintercept = -log10(0.05), colour = "black", linetype = "dashed", size = 0.25) + # Horizontal significance cut-off line.
  geom_vline(xintercept = 1, colour = "black", linetype = "dashed", size = 0.25)+  # Vertical significance cut-off line (+).
  geom_vline(xintercept = -1, colour = "black", linetype = "dashed", size = 0.25)  # Vertical significance cut-off line (+).
```

## S_TvsD

Down and up in Tolerant trees

```{r Plot individual S_TvsD , echo= FALSE}
ggplot(S_TvsD, aes(x=log2FoldChange_D2, y=sig_D2)) +
  geom_point(aes(colour =  color ),size =3.5)+
  scale_color_manual(values=c("#cd77ca", # purple D2 and ER
                              "#49cfd8", # blue Only D2
                              "#cdbd3a", # yellow Only ER
                              "grey",    # grey
                              "black"))+ # black
  xlab("log2 fold change")+
  ylab("-log10 (P value)")+
  theme_light(base_size = 10)+
  theme(legend.position = "none")+
  geom_hline(yintercept = -log10(0.05), colour = "black", linetype = "dashed", size = 0.25) + # Horizontal significance cut-off line.
  geom_vline(xintercept = 1, colour = "black", linetype = "dashed", size = 0.25)+  # Vertical significance cut-off line (+).
  geom_vline(xintercept = -1, colour = "black", linetype = "dashed", size = 0.25)  # Vertical significance cut-off line (+).
```
